requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10

pinecone-client==5.0.1
pinecone-text==0.9.0

openai==1.58.1

gdown==5.2.0